Update Core.swift
Removing the observers in `viewDidDisappear` leads to an issue with the interactive dismissal of `UIViewController`s.
When the interactive dismissal begins, `viewWillDisappear` is called and then, if the gesture in cancelled, `viewWillAppear` and `viewDidAppear` are called again but not `viewDidDisappear`.

This way, every time a dismissal is cancelled a new observer is added without removing the previous one. This results in multiple calls to `keyboardWillShow/Hide` when the keyboard is presented/dismissed.

Moving the `removeObserver` calls to `viewWillDisappear` should fix the issue.